Started work on spell toolbar + new command: change left click
Added a circle in the bottom left that indicates the current
left click spell.
M is now the command to change left click spells.
(Press M, then the net spell that you cast is set to the left click
command)

~This involved changes to the GUI file (all at the bottom)
and the hero file (unfortunately scattered since I changed
spell casting, added a boolean, and added a new key command)