Extended spellbar to include other spells - had issues though.
Unfortunately this is a little sloppy since I couldn't get the Gui file
to understand keycodes (I had to go online, find the corresponding
numbers, and manually type them in).
Is there a way to get this file to comprehend them, or even better, is
there a way to get the tags out of a map? The latter option would allow
a for-loop to take care of writing all the spell names, opposed to a
different function call for each one.

I'm slowing down on this until we figure out which class we want to
handle input events. (I assume the Gui as mentioned in the hero file)