Interact Function, and start of bank
Changed hero's collide function to just call nearby actors' interact
function when pressing down. The actor class has an empty version of
this function, with only the Pickupable and Portal classes currenlt
having their own versions of it.
Also added a enw bank tile, might be changed to an actor, not sure yet.